+++ backend/src/resolvers.ts
@@ -20,12 +20,12 @@ export const resolvers = {
         orderBy: Prisma.Enumerable<Prisma.PolicyOrderByWithRelationInput>;
         skip: number;
         take: number;
-        filter?: any;
+        filter: Status;
         search: string;
       },
       context: Context
     ): Promise<Policy[]> {
-      const where = args.search
+      const where: Prisma.PolicyWhereInput | undefined = args.search
         ? {
             OR: [
               { customer: { firstName: { contains: args.search } } },
