+++ frontend/src/helpers/headersArray.ts
@@ -1,6 +1,6 @@
-import { Policies } from "../interfaces/allPolicies.interface";
+import { Data } from "../interfaces/allPolicies.interface";
 
-export const getHeadersArray = (policyData: Policies) => {
+export const getHeadersArray = (policyData: Data) => {
   const headersArray = policyData?.allPolicies.reduce((acc: string[], curr) => {
     if (policyData.allPolicies.indexOf(curr) === 0) {
       return Object.keys(curr).slice(1);
