+++ frontend/src/interfaces/allPolicies.interface.ts
@@ -1,16 +1,20 @@
 // Generated by https://quicktype.io
 
-export interface Policies {
+export interface AllPolicies {
+  data: Data;
+}
+
+export interface Data {
   allPolicies: AllPolicy[];
   policiesCount: number;
 }
 
 export interface AllPolicy {
-  customer: Customer;
   provider: string;
+  customer: Customer;
   insuranceType: string;
   status: string;
-  policyNumber: string;
+  policyNumber: number;
   startDate: string;
   endDate: string;
   createdAt: string;
